backend/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Culturama is an application developed as part of the Final Course Work (TCC) of Etec de Taboão da Serra to promote cultural events."
authors = ["Giovanna Candido Manfredo <giovannamanfredo.dev@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.6"
django-environ = "^0.12.0"
validate-docbr = "^1.10.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
mysqlclient = "^2.2.7"
drf-spectacular = "^0.28.0"
pip = "^25.0.1"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.5.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
